Skip creating invoice if they already exist
